[buider] use builder() for cpu adam and fused optim in setup.py (#2187)
