+++ requirements/requirements.txt
@@ -7,3 +7,4 @@ rich
 click
 fabric
 contexttimer
+ninja
